pyproject.toml
[project]
name = "deliverant"
version = "0.1.0"
description = "Reliable webhook delivery platform"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery[redis]>=5.6.2",
    "cryptography>=46.0.3",
    "django>=6.0.1",
    "django-environ>=0.12.0",
    "django-prometheus>=2.4.0",
    "djangorestframework>=3.16.1",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "psycopg[binary]>=3.3.2",
    "python-json-logger>=4.0.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.11.1",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "deliverant.settings.development"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
testpaths = ["tests"]
addopts = "-v --tb=short"
